# api/pkg/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Echo Rest API
  description: A sample REST API
  termsOfService: https://example.com/terms-of-service
  license:
    name: TODO
    url: TODO
  version: 1.0.0
  contact:
    name: GDSC VIT
    url: https://github.com/GDGVIT
servers:
  - url: http://localhost:3000/v1
    description: Local development server
paths:
...
  "/vault/create":
    post:
      summary: Create a vault
      operationId: VaultCreate
      description: Use this endpoint to create a vault
      requestBody:
        description: Vault Creation Request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultCreateRequest"
      security:
        - UserJWT: []
      responses:
        "200":
          description: Vault Creation Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Vault Creation Response
                  value:
                    message: Vault was created successfully.
                error:
                  summary: Vault Creation Failure Response
                  value:
                    message: Vault could not be created.

  ## Get a vault
  "/vault/{pid}":
    get:
      summary: Get a vault by id
      operationId: VaultGet
      description: Use this endpoint to get a vault by id
      parameters:
        - in: path
          name: pid
          required: true
          description: Vault ID
          schema:
            type: string
      security:
        - UserJWT: []
      responses:
        "200":
          description: Vault Get Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultCreateRequest"
              examples:
                success:
                  summary: Successful Vault Get Response
                  value:
                    message: Vault was fetched successfully.
                error:
                  summary: Vault Get Failure Response
                  value:
                    message: Vault could not be fetched.

  ## Edit a vault
  "/vault/edit/{pid}":
    post:
      summary: Edit a vault
      operationId: VaultEdit
      description: Use this endpoint to edit a vault
      parameters:
        - in: path
          name: pid
          required: true
          description: Vault ID
          schema:
            type: string
      requestBody:
        description: Vault Edit Request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultEditRequest"
      security:
        - UserJWT: []
      responses:
        "200":
          description: Vault Edit Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Vault Edit Response
                  value:
                    message: Vault was edited successfully.
                error:
                  summary: Vault Edition Failure Response
                  value:
                    message: Vault could not be edited.
  ## Delete a vault
  "/vault/delete/{pid}":
    delete:
      summary: Delete a vault
      operationId: VaultDelete
      description: Use this endpoint to delete a vault
      parameters:
        - in: path
          name: pid
          required: true
          description: Vault ID
          schema:
            type: string
      security:
        - UserJWT: []
      responses:
        "200":
          description: Vault Delete Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessageResponse"
              examples:
                success:
                  summary: Successful Vault Delete Response
                  value:
                    message: Vault was deleted successfully.
                error:
                  summary: Vault Deletion Failure Response
                  value:
                    message: Vault could not be deleted.

  ##############################
  #### CREDENTIAL ENDPOINTS ####
  ##############################

  ## Create a credential
...
components:
  schemas:
...
    VaultCreateRequest:
      type: object
      required:
        - name
        - description
        - public_key
      properties:
        name:
          type: string
          example: Some Vault Name
        description:
          type: string
          example: Some Vault Description
        public_key:
          type: string
...
    VaultEditRequest:
      type: object
      properties:
        name:
          type: string
          example: Some Vault Name
        description:
          type: string
          example: Some Vault Description
        public_key:
          type: string
    GenericMessageResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
...
  securitySchemes:
    UserJWT:
      type: apiKey
      in: header
      name: Authorization